Adding booking and passenger controller

+++ MVCClient/Models/FlightM.cs
@@ -8,11 +8,16 @@ namespace MVCClient.Models
     public class FlightM
     {
         public int FlightNo { get; set; }
+
         public string Destination { get; set; }
+
         public DateTime Date { get; set; }
+
         public int Seats { get; set; }
+
         public int FreeSeats { get; set; }
-        public double FlightPrice { get; set; }
+
+        public int FlightPrice { get; set; }
 
     }
 }
