+++ WebAPIDoBeFly/Models/FlightM.cs
@@ -8,10 +8,15 @@ namespace WebAPIDoBeFly.Models
     public class FlightM
     {
         public int FlightNo { get; set; }
+
         public string Destination { get; set; }
+
         public DateTime Date { get; set; }
+
         public int Seats { get; set; }
+
         public int FreeSeats { get; set; }
-        public double FlightPrice { get; set; }
+
+        public int FlightPrice { get; set; }
     }
 }
